fix: 디코딩에러 수정
